+++ src/example_app/example_app/documents/schemas.py
@@ -3,19 +3,22 @@
 :Authors: cykooz
 :Date: 31.07.2020
 """
+
 import decimal
 
 import colander
 from restfw import schemas
 
 from restfw_admin import widgets
+from restfw_admin.schemas import FileNode
 from ..users.schemas import user_id_validator
 
 
 class DocMetaDataSchema(schemas.MappingNode):
     type = schemas.EmptyStringNode(title='Document type')
     custom = schemas.MappingNode(
-        title='Custom data', unknown='preserve',
+        title='Custom data',
+        unknown='preserve',
         widget=widgets.JsonField(),
     )
 
@@ -23,7 +26,8 @@ class DocMetaDataSchema(schemas.MappingNode):
 class EditDocMetaDataSchema(schemas.MappingNode):
     type = schemas.EmptyStringNode(title='Document type', missing='')
     custom = schemas.MappingNode(
-        title='Custom data', unknown='preserve',
+        title='Custom data',
+        unknown='preserve',
         widget=widgets.JsonInput(
             initial_value={},
             full_width=True,
@@ -36,8 +40,10 @@ class DocSchema(schemas.HalResourceSchema):
     user_id = schemas.IntegerNode(title='User ID')
     name = schemas.StringNode(title='Document name')
     data = schemas.EmptyStringNode(title='Document data')
+    image = FileNode(title='Document image', nullable=True)
     publish_date = schemas.DateTimeNode(
-        title='Publish date', nullable=True,
+        title='Publish date',
+        nullable=True,
     )
     weight = colander.SchemaNode(
         colander.Decimal('.00', decimal.ROUND_HALF_UP),
@@ -51,7 +57,8 @@ class DocsSchema(schemas.HalResourceWithEmbeddedSchema):
     _embedded = schemas.EmbeddedNode(
         schemas.SequenceNode(
             DocSchema(title='Document'),
-            name='docs', title='List of embedded documents'
+            name='docs',
+            title='List of embedded documents',
         ),
         missing=colander.drop,
     )
@@ -59,7 +66,8 @@ class DocsSchema(schemas.HalResourceWithEmbeddedSchema):
 
 class GetDocsSchema(schemas.GetEmbeddedSchema):
     name = schemas.StringNode(
-        title='Document name', missing=colander.drop,
+        title='Document name',
+        missing=colander.drop,
     )
 
 
@@ -67,9 +75,11 @@ class CreateDocSchema(schemas.MappingNode):
     user_id = schemas.IntegerNode(title='User ID', validator=user_id_validator)
     name = schemas.StringNode(title='Document name')
     data = schemas.EmptyStringNode(title='Document data')
+    image = FileNode(title='Document image', nullable=True)
     meta = EditDocMetaDataSchema(title='Meta data')
     publish_date = schemas.DateTimeNode(
-        title='Publish date', nullable=True,
+        title='Publish date',
+        nullable=True,
         missing=None,
     )
     weight = colander.SchemaNode(
@@ -81,7 +91,8 @@ class CreateDocSchema(schemas.MappingNode):
 
 
 PatchDocSchema = schemas.clone_schema_class(
-    'PatchDocSchema', CreateDocSchema,
+    'PatchDocSchema',
+    CreateDocSchema,
     excludes=['user_id'],
     nodes_missing=colander.drop,
 )