+++ src/example_app/example_app/documents/views.py
@@ -2,46 +2,51 @@
 :Authors: cykooz
 :Date: 14.01.2021
 """
+
 from dataclasses import asdict
 
 from restfw import views
 from restfw.interfaces import MethodOptions
-from restfw.schemas import GetEmbeddedSchema
 
 from . import schemas
-from .resources import Doc, Docs
+from .resources import Doc, Docs, File
 
 
 # Doc
 
+
 @views.resource_view_config(Doc)
 class DocView(views.HalResourceView):
     resource: Doc
-    options_for_get = MethodOptions(
-        None, schemas.DocSchema,
-        permission='docs.get'
-    )
+    options_for_get = MethodOptions(None, schemas.DocSchema, permission='docs.get')
     options_for_patch = MethodOptions(
-        schemas.PatchDocSchema, schemas.DocSchema,
+        schemas.PatchDocSchema,
+        schemas.DocSchema,
         permission='docs.edit',
     )
     options_for_delete = MethodOptions(None, None, permission='docs.edit')
 
     def as_dict(self):
-        return asdict(self.resource.model)
+        res = asdict(self.resource.model)
+        if res['image'] is None:
+            res['image'] = asdict(File())
+        return res
 
 
 # Docs
+
 
 @views.resource_view_config(Docs)
 class DocsView(views.HalResourceWithEmbeddedView):
     resource: Docs
     options_for_get = MethodOptions(
-        schemas.GetDocsSchema, schemas.DocsSchema,
+        schemas.GetDocsSchema,
+        schemas.DocsSchema,
         permission='docs.get',
     )
     options_for_post = MethodOptions(
-        schemas.CreateDocSchema, schemas.DocSchema,
+        schemas.CreateDocSchema,
+        schemas.DocSchema,
         permission='docs.edit',
     )
 
@@ -53,7 +58,9 @@ class DocsView(views.HalResourceWithEmbeddedView):
             if not name or model.name == name
         ]
         return views.list_to_embedded_resources(
-            self.request, params, docs,
+            self.request,
+            params,
+            docs,
             parent=self.resource,
             embedded_name='docs',
         )