+++ src/restfw_admin/CHANGES.rst
@@ -14,6 +14,15 @@
 CHANGELOG
 *********
 
+Next release
+============
+
+Features
+--------
+
+- Added ``FileField`` and ``FileInput`` widgets.
+- Added custom implementation of ``IFileUploadProvider``.
+
 1.4 (2024-12-11)
 ================